+++ src/components/ui/command.tsx
@@ -76,6 +76,7 @@ function CommandInput({
                     "placeholder:text-muted-foreground flex h-10 w-full rounded-md bg-transparent py-3 text-sm outline-hidden disabled:cursor-not-allowed disabled:opacity-50",
                     className
                 )}
+                style={{ outline: "none" }}
                 {...props}
             />
         </div>